refactor(cache-schema): split writer and std impls
